requirements.txt
streamlit
groq
python-dotenv
google-generativeai
streamlit
groq
python-dotenv
PyYAML
google-generativeai
speechrecognition
fpdf
altair
pandas
PyPDF2
python-docx
streamlit-webrtc
requests
streamlit
gtts
streamlit-webrtc
requests
PyYAML
fpdf
groq
python-dotenv
